normalizationUpdateJobQueue
