--- kifi-backend/modules/common/app/com/keepit/common/queue/NormalizationUpdateJobQueue.scala
@@ -1,6 +1,0 @@
-package com.keepit.common.queue
-
-
-trait NormalizationUpdateJobQueue extends SimpleQueue {
-  override val name = "NTest"
-}
