+++ kifi-backend/modules/shoebox/app/com/keepit/dev/ShoeboxDevModule.scala
@@ -13,6 +13,7 @@ import com.keepit.inject.CommonDevModule
 import com.keepit.common.integration.DevReaperModule
 import com.keepit.scraper.DevScrapeSchedulerModule
 import com.keepit.common.queue.{ProdSimpleQueueModule, DevSimpleQueueModule}
+import com.keepit.queue.DevNormalizationUpdateJobQueueModule
 
 case class ShoeboxDevModule() extends ShoeboxModule(
   secureSocialModule = ProdShoeboxSecureSocialModule(),
@@ -20,6 +21,7 @@ case class ShoeboxDevModule() extends ShoeboxModule(
   reaperModule = DevReaperModule(),
   storeModule = ShoeboxDevStoreModule(),
   sqsModule = DevSimpleQueueModule(),
+  normalizationQueueModule = DevNormalizationUpdateJobQueueModule(),
 
   // Shoebox Functional Modules
   analyticsModule = DevAnalyticsModule(),
