+++ kifi-backend/modules/shoebox/test/com/keepit/queue/FakeNormalizationUpdateJobQueueModule.scala
@@ -1,0 +1,20 @@
+package com.keepit.queue
+
+import com.google.inject.{Provides, Singleton}
+import com.keepit.common.queue.SQSMessage
+
+case class FakeNormalizationUpdateJobQueueModule() extends NormalizationUpdateJobQueueModule {
+
+  override def configure(): Unit = {
+  }
+
+  @Singleton
+  @Provides
+  def normalizationUpdateTaskQ():NormalizationUpdateJobQueue = new NormalizationUpdateJobQueue {
+    override def queueUrl: String = "http://foo.com/barQ"
+    override def receive(): Seq[SQSMessage] = Seq.empty[SQSMessage]
+    override def send(s: String): Unit = {}
+    override def delete(msgHandle: String): Unit = {}
+  }
+
+}
